+++ test/spec/Feature/Query/PreferencesSpec.hs
@@ -192,16 +192,25 @@ spec =
           { matchStatus = 204
           , matchHeaders = ["Preference-Applied" <:> "handling=lenient"]}
 
-      it "should fail with rpc when deleting rows more than prefered" $
-        request methodPost "/rpc/delete_all_items"
+    context "test Prefer: max-affected with rpc" $ do
+      it "should fail with rpc when deleting rows more than prefered with returns setof" $
+        request methodPost "/rpc/delete_items_returns_setof"
           [("Prefer", "handling=strict, max-affected=10")]
           ""
           `shouldRespondWith`
           [json| {"code":"PGRST124","details":"The query affects 15 rows","hint":null,"message":"Query result exceeds max-affected preference constraint"} |]
           { matchStatus = 400 }
 
-      it "should succeed with rpc deleting rows less than prefered" $
-        request methodPost "/rpc/delete_all_items"
+      it "should fail with rpc when deleting rows more than prefered with returns table" $
+        request methodPost "/rpc/delete_items_returns_table"
+          [("Prefer", "handling=strict, max-affected=10")]
+          ""
+          `shouldRespondWith`
+          [json| {"code":"PGRST124","details":"The query affects 15 rows","hint":null,"message":"Query result exceeds max-affected preference constraint"} |]
+          { matchStatus = 400 }
+
+      it "should succeed with rpc deleting rows less than prefered with returns setof" $
+        request methodPost "/rpc/delete_items_returns_setof"
           [("Prefer", "handling=strict, max-affected=20")]
           ""
           `shouldRespondWith`
@@ -209,3 +218,21 @@ spec =
                  {"id":8},{"id":9},{"id":10},{"id":11},{"id":12},{"id":13},
                  {"id":14},{"id":15}]|]
           { matchStatus = 200 }
+
+      it "should succeed with rpc deleting rows less than prefered with returns table" $
+        request methodPost "/rpc/delete_items_returns_table"
+          [("Prefer", "handling=strict, max-affected=20")]
+          ""
+          `shouldRespondWith`
+          [json|[{"id":1},{"id":2},{"id":3},{"id":4},{"id":5},{"id":6},{"id":7},
+                 {"id":8},{"id":9},{"id":10},{"id":11},{"id":12},{"id":13},
+                 {"id":14},{"id":15}]|]
+          { matchStatus = 200 }
+
+      it "should fail with rpc when returns void with handling=strict" $
+        request methodPost "/rpc/delete_items_returns_void"
+          [("Prefer", "handling=strict, max-affected=20")]
+          ""
+          `shouldRespondWith`
+          [json| {"code":"PGRST128","details":null,"hint":null,"message":"Function must return SETOF or TABLE when max-affected preference is used with handling=strict"} |]
+          { matchStatus = 400 }